@@ -16,30 +16,30 @@ class AnimSearchBar extends StatefulWidget {
 
   final double width;
   final TextEditingController textController;
-  final Icon suffixIcon;
-  final Icon prefixIcon;
+  final Icon? suffixIcon;
+  final Icon? prefixIcon;
   final String helpText;
   final int animationDurationInMilli;
   final onSuffixTap;
   final bool rtl;
   final bool autoFocus;
-  final TextStyle style;
+  final TextStyle? style;
   final bool closeSearchOnSuffixTap;
 
   const AnimSearchBar({
-    Key key,
+    Key? key,
 
     /// The width cannot be null
-    @required this.width,
+    required this.width,
 
     /// The textController cannot be null
-    @required this.textController,
+    required this.textController,
     this.suffixIcon,
     this.prefixIcon,
     this.helpText = "Search...",
 
     /// The onSuffixTap cannot be null
-    @required this.onSuffixTap,
+    required this.onSuffixTap,
     this.animationDurationInMilli = 375,
 
     /// make the search bar to open from right to left
@@ -72,7 +72,7 @@ int toggle = 0;
 class _AnimSearchBarState extends State<AnimSearchBar>
     with SingleTickerProviderStateMixin {
   ///initializing the AnimationController
-  AnimationController _con;
+  late AnimationController _con;
   FocusNode focusNode = FocusNode();
   @override
   void initState() {
@@ -227,7 +227,7 @@ class _AnimSearchBarState extends State<AnimSearchBar>
                 icon: widget.prefixIcon != null
                     ? toggle == 1
                         ? Icon(Icons.arrow_back_ios)
-                        : widget.prefixIcon
+                        : widget.prefixIcon!
                     : Icon(
                         toggle == 1 ? Icons.arrow_back_ios : Icons.search,
                         size: 20.0,
